xhs-insight_v1.5.1/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.5.0
python-multipart>=0.0.6
requests>=2.31.0
google-genai>=0.2.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
httpx>=0.25.0
PyExecJS
loguru
python-dotenv
retry
openpyxl